requirements.txt

python-telegram-bot==13.15
schedule==1.1.0
